pyproject.toml
[tool.poetry]
name = "langchain-study"
version = "0.1.0"
description = ""
authors = ["jurepi <danhee.an@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
scipy = "^1.13.1"
python-dotenv = "^1.0.1"
openai = "^1.30.5"
playwright = "^1.44.0"
beautifulsoup4 = "^4.12.3"
nest-asyncio = "^1.6.0"
html2text = "^2024.2.26"
langchain-openai = "0.1.7"
langchain = "0.2.3"
langchain-community = "^0.2.1"
langchain-chroma = "^0.1.1"
pygments = "^2.18.0"
google-api-python-client = "^2.131.0"
apify-client = "^1.7.0"
langserve = "^0.2.2"
sse-starlette = "^2.1.0"
chromadb = "0.4.1"
langgraph = "^0.0.68"
langchain-upstage = "^0.1.6"
streamlit = "^1.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
